+++ dlib/lzp_buffer/lzp_buffer_kernel_2.h
@@ -63,7 +63,7 @@ namespace dlib
             unsigned long& index
         );
 
-        inline unsigned long size (
+        inline size_t size (
         ) const;
 
         inline unsigned char operator[] (
@@ -291,7 +291,7 @@ namespace dlib
     template <
         typename sbuf
         >
-    unsigned long lzp_buffer_kernel_2<sbuf>::
+    size_t lzp_buffer_kernel_2<sbuf>::
     size (
     ) const 
     {
