+++ dlib/sequence/sequence_kernel_1.h
@@ -144,7 +144,7 @@ namespace dlib
             );
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             bool at_start (
@@ -446,7 +446,7 @@ namespace dlib
         typename T,
         typename mem_manager
         >
-    unsigned long sequence_kernel_1<T,mem_manager>::
+    size_t sequence_kernel_1<T,mem_manager>::
     size (
     ) const
     {
